Extract two functions for readability and testability
In compare_fst, two transducers are read and compared; the reader code was
written twice (except for the prints saying "old" and "new"), I have moved it
into a function read_transducer.

I have also moved process_row out of compare_fst (could be _process_row if we
really want to communicate its local use) to make it testable (later...), make
it simpler (it is now obvious that it doesn't rely on a closure to work)  and
reduce the size of the compare_fst code (still 250 lines!).