Handle first the case where transducer != 'internal'
Arguably, this isn't a very useful change, but it highlights the fact that
new_transducer is never an empty string. I currently feel unhappy about
this section of the function: there is a bit too much magic around the
whole "if internal, then use pipeline_name but if pipeline_name cannot be
found, use refish", but I think cleaning it requires changes a bit more
than locally, so moving on.