/!\ Fix Germanic FST
(it's only a fix if I didn't screw everything up...)
There was a line in the code that said that if the pipeline is Germanic, then
process the syllables differently from the other cases. My understanding is
that this was due to the Germanic FST having weird definitions like "a " (with
a space afterwards). This is strange enough that it looks like it's on purpose,
but I don't know the purpose (or if it's because of a previous state of the
code base); when removing those extra spaces, then the FST works without the
special handling of Germanic, so I went ahead.